fix: make ComponentTypeRegistry Engine-scoped and use proper cloning
- Remove global singleton pattern from ComponentTypeRegistry
- Each ArchetypeManager now owns its own ComponentTypeRegistry instance
- Pass registry to Archetype constructor for proper scoping
- This allows multiple independent Engine instances

Also fixes:
- Replace JSON.parse/stringify cloning with deepCloneComponent utility
- Move deepCloneComponent to utils.ts for reuse across modules
- Export deepCloneComponent and ComponentTypeRegistry from public API
- Update archetype.spec.ts to use registry parameter

Fixes architecture issue where global singleton prevented independent engines.
